Fix about 12 kinds of warnings over the codebase

Commit migrated from https://github.com/dotnet/coreclr/commit/465721555eda899a06bc91fe37bc47c0a1adbdd4

+++ src/clr/src/debug/daccess/nidump.cpp
@@ -2836,7 +2836,7 @@ IMetaDataImport2 * NativeImageDumper::TypeToString(PTR_CCOR_SIGNATURE &sig,
     {
         buf.AppendPrintf( "%s", elementNames[type] );
     }
-    else switch (type)
+    else switch ((DWORD)type)
     {
     case ELEMENT_TYPE_CANON_ZAPSIG:
         buf.Append( W("System.__Canon") );
